meeting_organizer/project_backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
django-cors-headers==4.3.1
django-ninja==1.1.0
pydantic==2.6.2
pydantic-core==2.16.3
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
